dataset_generation_scripts/requirements.txt
alabaster==0.7.12
anyio==2.2.0
argon2-cffi==20.1.0
arrow==1.2.2
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
binaryornot==0.4.4
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.3
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
docutils==0.16
docx2txt==0.8
entrypoints==0.3
idna==2.10
imagesize==1.2.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.1.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.11.1
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.6.7
notebook==6.3.0
npzviewer==0.2.0
numpy==1.20.1
open3d==0.14.1
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==9.0.1
pip-chill==1.0.1
poyo==0.5.0
prometheus-client==0.9.0
prompt-toolkit==3.0.18
pycparser==2.20
Pygments==2.7.4
pymeshlab==2021.10
pyntcloud==0.1.6
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-slugify==6.1.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
regex==2021.11.10
requests==2.25.1
scikit-learn==1.1.0
scipy==1.8.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
snowballstemmer==2.1.0
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
terminado==0.9.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==3.1.0
tomlkit==0.9.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zstandard==0.15.2
